Separate logging from metrics
- Move all logging telemetry hooks to `NervesHub.Logger`
- Separate polling and statsd setup
- Remove the need for a supervisor, just use `NervesHub.Application`
- Only setup the statsd reporter if the config exists
