+++ lib/nerves_hub/logger.ex
@@ -4,7 +4,11 @@ defmodule NervesHub.Logger do
   @doc false
   def install() do
     handlers = %{
-      [:phoenix, :endpoint, :stop] => &__MODULE__.phoenix_endpoint_stop/4
+      [:phoenix, :endpoint, :stop] => &__MODULE__.log_event/4,
+      [:nerves_hub, :devices, :connect] => &__MODULE__.log_event/4,
+      [:nerves_hub, :devices, :disconnect] => &__MODULE__.log_event/4,
+      [:nerves_hub, :devices, :duplicate_connection] => &__MODULE__.log_event/4,
+      [:nerves_hub, :devices, :update, :automatic] => &__MODULE__.log_event/4
     }
 
     for {key, fun} <- handlers do
@@ -17,7 +21,7 @@ defmodule NervesHub.Logger do
   # Phoenix request logging
 
   @doc false
-  def phoenix_endpoint_stop(_, %{duration: duration}, %{conn: conn} = metadata, _) do
+  def log_event([:phoenix, :endpoint, :stop], %{duration: duration}, %{conn: conn} = metadata, _) do
     case log_level(metadata[:options][:log], conn) do
       false ->
         :ok
@@ -33,6 +37,47 @@ defmodule NervesHub.Logger do
           )
         end)
     end
+  end
+
+  def log_event([:nerves_hub, :devices, :connect], _, metadata, _) do
+    Logger.info("Device connected",
+      event: "nerves_hub.devices.connect",
+      identifier: metadata[:identifier],
+      firmware_uuid: metadata[:firmware_uuid]
+    )
+  end
+
+  def log_event([:nerves_hub, :devices, :duplicate_connection], _, metadata, _) do
+    Logger.info("Device duplicate connection detected",
+      event: "nerves_hub.devices.duplicate_connection",
+      ref_id: metadata[:ref_id],
+      identifier: metadata[:device].identifier
+    )
+  end
+
+  def log_event([:nerves_hub, :devices, :disconnect], _, metadata, _) do
+    Logger.info("Device disconnected",
+      event: "nerves_hub.devices.disconnect",
+      ref_id: metadata[:ref_id],
+      identifier: metadata[:identifier]
+    )
+  end
+
+  def log_event([:nerves_hub, :devices, :update, :automatic], _, metadata, _) do
+    Logger.info("Device received update",
+      event: "nerves_hub.devices.update.automatic",
+      ref_id: metadata[:ref_id],
+      identifier: metadata[:identifier],
+      firmware_uuid: metadata[:firmware_uuid]
+    )
+  end
+
+  def log_event([:nerves_hub, :devices, :update, :successful], _, metadata, _) do
+    Logger.info("Device updated firmware",
+      event: "nerves_hub.devices.update.successful",
+      identifier: metadata[:identifier],
+      firmware_uuid: metadata[:firmware_uuid]
+    )
   end
 
   # Helper functions
